+++ tests/Exceptions/E4xx_FilesystemExceptionTest.php
@@ -34,46 +34,58 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Filesystem/Checks
+ * @package   Filesystem/Exceptions
  * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
  * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://code.ganbarodigital.com/php-file-system
  */
 
-namespace GanbaroDigital\Filesystem\Checks;
+namespace GanbaroDigital\Filesystem\Exceptions;
 
-use GanbaroDigital\Filesystem\DataTypes\FilesystemPathData;
+use PHPUnit_Framework_TestCase;
+use RuntimeException;
 
-class BaseFilenameCheck implements FilesystemPathCheck
+/**
+ * @coversDefaultClass GanbaroDigital\Filesystem\Exceptions\E4xx_FilesystemException
+ */
+class E4xx_FilesystemExceptionTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * does the provided filesystem path data meet the requirements for this
-     * specification?
-     *
-     * @param  FilesystemPathData $fsData
-     *         the data to inspect
-     * @return boolean
-     *         TRUE if the filesystem data meets the requirements
-     *         FALSE otherwise
+     * @covers ::__construct
      */
-    public static function checkFilesystemPathData(FilesystemPathData $fsData)
+    public function testCanInstantiate()
     {
-        $path = (string)$fsData;
-        return static::checkString($path);
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $expectedCode = 100;
+        $expectedMessage = "hello cruel world";
+
+        $obj = new E4xx_FilesystemException($expectedCode, $expectedMessage);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($obj instanceof E4xx_FilesystemException);
     }
 
     /**
-     * does the provided filename meet the requirements for this specification?
-     *
-     * @param  string $path
-     *         the filename to check
-     * @return boolean
-     *         TRUE if the filename meets the requirements
-     *         FALSE otherwise
+     * @covers ::__construct
      */
-    public static function checkString($path)
+    public function testExtendsExxx_FilesystemException()
     {
-        return false;
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $expectedCode = 100;
+        $expectedMessage = "hello cruel world";
+
+        $obj = new E4xx_FilesystemException($expectedCode, $expectedMessage);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($obj instanceof Exxx_FilesystemException);
     }
 }