+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 Nothing yet.
 
+## 0.2.0 - Mon Sep 21 2015
+
+### Fixes
+
+* Checks - now all support the standard `::check()` method
+* Checks - dropped the interface for now (needs rethinking)
+* Checks - stricter type checking
+* E4xx_UnsupportedType - typo in `use` statement
+
+### Test Fixes
+
+* Added `Exceptions\Exxx_FilesystemException`
+* Added `Exceptions\E4xx_FilesystemException`
+* Added `Exceptions\E4xx_UnsupportedType`
+
 ## 0.1.0 - Wed Sep 2 2015
 
 Initial release.